feat: rename to CustomEmailValidator_server.ts camelcase. remove JSON.parse bc already a JS object - not JSON. add EmailController for POST email/send route
